requirements.txt
alphabet-detector==0.0.7
asgiref==3.7.2
aws-psycopg2==1.3.8
boto==2.49.0
boto3==1.29.3
botocore==1.32.3
certifi==2023.11.17
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.7
django-autofixture==0.12.1
django-boto==0.3.12
django-cors-headers==4.3.1
django-filter==23.3
django-modeltranslation==0.18.11
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-seed==0.3.1
django-storages==1.14.2
djangorestframework==3.14.0
drf-yasg==1.21.7
Faker==22.5.1
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.3
jmespath==1.0.1
jsonfield==3.1.0
MarkupSafe==2.1.4
openapi-codec==1.3.2
packaging==23.2
Pillow==10.1.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.7.0
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.7
